Made it so long_first_name and long_last_name columns aren't made if they are not needed.
Also changed how merge works a bit so it doesn't matter where
first_name, last_name, and email columns are since it just ignores them.
It also now merges on email then merges on contact_id to help reduce
duplicates even more if columns are in the df. It is also easier now to
merge on other things if needed.